Allow redirects to external websites
Removing a forward slash at the location header will allow you to specify external domains (e.g. https://example.org/) in the JSON file. However, with this change there would be no backwards compatibility with the master (original), considering that then a redirects.json file has to be modified to have the redirect value start with a forward slash (/).
To combat this I added a section which would check if it's an internal redirect without a forward slash at the beginning, and if so it will add a / at the beginning.